+++ crates/core/src/manifest.rs
@@ -51,7 +51,7 @@ impl Manifest {
     pub fn load<P: AsRef<Path>>(path: P) -> Result<Self, ProtoError> {
         let path = path.as_ref();
 
-        debug!(file = %path.display(), "Loading manifest");
+        debug!(file = ?path, "Loading manifest");
 
         let mut manifest: Manifest = if path.exists() {
             use fs4::FileExt;
@@ -90,7 +90,7 @@ impl Manifest {
         use fs4::FileExt;
         use std::io::{prelude::*, SeekFrom};
 
-        debug!(file = %self.path.display(), "Saving manifest");
+        debug!(file = ?self.path, "Saving manifest");
 
         if let Some(parent) = self.path.parent() {
             fs::create_dir_all(parent)?;